pyproject.toml
[project]
name = "fixfanbox"
version = "0.1.0"
description = "Add your description here"
authors = [
    {name = "xtaodada", email = "xtao@xtao.de"},
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "coloredlogs>=15.0.1",
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "jinja2>=3.1.5",
    "lxml>=5.3.1",
    "persica>=0.1.0a5",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.0",
    "python-dotenv>=1.0.1",
    "starlette>=0.45.3",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "ruff>=0.9.7",
]
